rename pkg to authdb
